+++ hooks/index.js
@@ -1,3 +1,0 @@
-export { default as useLocationsForTenants } from './useLocationsForTenants';
-export { default as useInstancesQuery } from './useInstancesQuery';
-export * from './useConsortiaTenants';