+++ index.js
@@ -1,4 +1,1 @@
-import React from 'react';
-import InstanceSearch from './InstanceSearch';
-
-export default props => <InstanceSearch {...props} />;
+export { default } from './src';