+++ src/components/index.js
@@ -1,0 +1,3 @@
+export * from './FindInstanceContainer';
+export * from './PluginFindRecord';
+export * from './FilterNavigation';