+++ test/jest/__mock__/stripesIcon.mock.js
@@ -1,0 +1,5 @@
+jest.mock('@folio/stripes-components/lib/Icon', () => {
+  return jest.fn(props => {
+    return props.children ? props.children : <span />;
+  });
+});