+++ test/jest/__mock__/stripesSmartComponents.mock.js
@@ -1,96 +1,34 @@
 import React from 'react';
 
 jest.mock('@folio/stripes/smart-components', () => ({
-  makeQueryFunction: jest.fn(() => jest.fn()),
-  SearchAndSortQuery: jest.fn(({ children, ...rest }) => <div {...rest}>{children}</div>),
-  SearchAndSortNoResultsMessage: jest.fn(({
-    label,
-    filterPaneIsVisible = true,
-    toggleFilterPane = jest.fn(),
-    ...rest
-  }) => (
-    <div {...rest}>
+  ...jest.requireActual('@folio/stripes/smart-components'),
+  LocationLookup: () => <div>LocationLookup</div>,
+  ViewMetaData: () => <div>ViewMetaData</div>,
+  ControlledVocab: jest.fn(() => <div>ControlledVocab</div>),
+  ConfigManager: (props) => {
+    const { getInitialValues, onBeforeSave, children } = props;
+    const component =
       <div>
-        <span>{label}</span>
-      </div>
-      {!filterPaneIsVisible &&
-        <button
-          type="submit"
-          onClick={toggleFilterPane}
-        >
-          Show filters
-        </button>
+        <div>ConfigManager</div>
+        {children}
+        <button type="button" onClick={() => getInitialValues()}>getInitialValues</button>
+        <button type="button" onClick={() => onBeforeSave()}>onBeforeSave</button>
+      </div>;
+    return component;
+  },
+  useRemoteStorageMappings: () => {
+    return ({
+      'holdings-id-1': {
+        'id': 'holdings-id-1',
+        'name': 'Storage A',
+        'description': 'Storage A description'
+      },
+      'holdings-id-2': {
+        'id': 'holdings-id-2',
+        'name': 'Storage B',
+        'description': 'Storage B description'
       }
-    </div>
-  )),
-  MultiSelectionFilter: jest.fn(({
-    name,
-    onChange = jest.fn()
-  }) => (
-    <div>
-      <div>
-        <span>{name}</span>
-      </div>
-      <input
-        type="text"
-        onChange={onChange()}
-        data-testid={`${name}`}
-      />
-    </div>
-  )),
-  DateRangeFilter: jest.fn(({
-    name,
-    onChange = jest.fn(),
-  }) => (
-    <div>
-      <div>
-        <span>{name}</span>
-      </div>
-      <button
-        type="submit"
-        onClick={onChange}
-      >
-        DateRangeFilter
-      </button>
-    </div>
-  )),
-  CheckboxFilter: jest.fn(({
-    name,
-    onChange = jest.fn(),
-  }) => (
-    <div>
-      <div>
-        <span>{name}</span>
-      </div>
-      <button
-        type="submit"
-        onClick={onChange}
-      >
-        CheckboxFilter
-      </button>
-    </div>
-  )),
-  SearchAndSortSearchButton: jest.fn(({
-    label,
-    id,
-    onClick = jest.fn(),
-    disabled,
-    ...restProps
-  }) => (
-    <div>
-      <button
-        type="button"
-        buttonStyle="none"
-        id={id}
-        onClick={onClick}
-        disabled={disabled}
-        {...restProps}
-      >
-        <span size="small" icon="times-circle-solid">
-          {label}
-        </span>
-      </button>
-    </div>
-  )),
-  StripesConnectedSource: jest.fn()
+    });
+  }
 }), { virtual: true });
+